Split train code into own script
Lightning will start scripts with different environment variables when
using its `ddp` strategy. It'll do so by calling `python
script_entrypoint.py` where `script.py` is the entrypoint for the
program.

This doesn't work with Modal because our entrypoint is
_container_entrypoint.py, a Modal Python program executes a function
deserializing its inputs from our backend. Lightning will attempt to run
_container_entrypoint.py but it will fail.

We can opt to use `ddp_spawn` in this case (or `ddp_notebook`) but all
program objects need to be pickable. It looks like some aren't, so that
didn't work without more debugging.

The solution then is to isolate the training code in its own script,
then call it from a Modal function with subprocess. Lightning will do
the right thing then, stating one process per GPU by calling the
training script. The Modal program will only observe the subprocess.
